fixed shots now animating again. Added enemy red color when hit.

+++ app/src/main/java/com/gamelogic/Shooter.java
@@ -26,7 +26,7 @@ import static com.graphics.GraphicsTools.LL;
 
 
 public class Shooter {
-    private static final int BASE_SPEED = 20;
+    private static final int BASE_SPEED = 30;
     private SharedPreferences preferences;
     private AudioPlayer audioPlayer;
     private WeaponsHandler weaponsHandler;
@@ -39,11 +39,10 @@ public class Shooter {
 
         public void move() {
             direction.set(direction.getAngle(), BASE_SPEED);
-            shot.placeAt(500, 500);
-            shot.moveBy(DataContainer.mapMovement.x, DataContainer.mapMovement.y, 0);
+            shot.moveBy(-DataContainer.mapMovement.x, -DataContainer.mapMovement.y, 0);
             shot.move(direction);
-            shot.getPosition().x += direction.velocity_X;
-            shot.getPosition().y += direction.velocity_Y;
+//            shot.getPosition().x += direction.velocity_X;
+//            shot.getPosition().y += direction.velocity_Y;
         }
     }
 
@@ -52,7 +51,7 @@ public class Shooter {
     private Direction baseDirection;
 
     public Shooter(WeaponsHandler weaponsHandler) {
-        this.weaponsHandler =  weaponsHandler;
+        this.weaponsHandler = weaponsHandler;
 
         shotFactory = new SpriteEntityFactory(R.drawable.bullets, 20, 30, 1, 1, new PointF(400, 400));
         shots = new ArrayList<>();
@@ -61,6 +60,8 @@ public class Shooter {
     }
 
     public void shoot(PointF shooterGlobalPos, RectF shooterBaseRect, Direction shooterDirection, WeaponsHandler.weaponList_e currentWeapon) {
+
+
         if (weaponsHandler.getCurrentAmmoAmount() > 0) {
             LL(this, "crating a shot " +
 
@@ -69,32 +70,32 @@ public class Shooter {
                     " shooterDirection " + shooterDirection.toString()
             );
 
-            if (preferences.getBoolean("sound", true)) {
-                switch (currentWeapon) {
-                    case GUN:
-                        audioPlayer.playAudioFromRaw(R.raw.gun);
-                        break;
-                    case SHOTGUN:
-                        audioPlayer.playAudioFromRaw(R.raw.shotgun);
-                        break;
-                    case AK47:
-                        audioPlayer.playAudioFromRaw(R.raw.ak);
-                        break;
-                    default:
-                        Log.e("SHOOTER", "DEFAULTED IN SHOOTER::shoot: switch (currentWeapon)");
-                        break;
-                }
-            }
+//            if (preferences.getBoolean("sound", true)) {
+//                switch (currentWeapon) {
+//                    case GUN:
+//                        audioPlayer.playAudioFromRaw(R.raw.gun);
+//                        break;
+//                    case SHOTGUN:
+//                        audioPlayer.playAudioFromRaw(R.raw.shotgun);
+//                        break;
+//                    case AK47:
+//                        audioPlayer.playAudioFromRaw(R.raw.ak);
+//                        break;
+//                    default:
+//                        Log.e("SHOOTER", "DEFAULTED IN SHOOTER::shoot: switch (currentWeapon)");
+//                        break;
+//                }
+//            }
 
             Shot s = new Shot();
             s.shot = shotFactory.createEntity();
             s.shot.placeAt(shooterBaseRect.centerX(), shooterBaseRect.centerY());
             s.shot.setPosition(shooterGlobalPos);
             s.direction = new Direction(shooterDirection, 1);
-            s.shot.setCurrentSprite(0);
-            s.shot.setAngleOffSet(0);
+//            s.shot.setCurrentSprite(0);
+//            s.shot.setAngleOffSet(0);
             shots.add(s);
-            weaponsHandler.setCurrentAmmoAmount(weaponsHandler.getCurrentAmmoAmount() - 1);
+//            weaponsHandler.setCurrentAmmoAmount(weaponsHandler.getCurrentAmmoAmount() - 1);
             Log.d("SHOOTER", "AMMO LEFT: " + weaponsHandler.getCurrentAmmoAmount());
         } else {
             audioPlayer.playAudioFromRaw(R.raw.dry_fire);
@@ -109,20 +110,23 @@ public class Shooter {
 
         for (Iterator<Shot> it = shots.iterator(); it.hasNext(); ) {
             Shot s = it.next();
-            LL(this, "Shoter update at: " + Arrays.toString(GraphicsTools.getCornersFromRect(s.shot.getRect())));
-
+//            LL(this, "Shoter update at: " + Arrays.toString(GraphicsTools.getCornersFromRect(s.shot.getRect())));
+//
             if (s.animationCounter < 5) {
                 s.animationCounter++;
             } else {
-                s.shot.setCurrentSprite(1);
+//                s.shot.setCurrentSprite(1);
             }
             s.move();
 
-            if (!DataContainer.mapBaseRect.contains(s.shot.getRect())) {
+            if (s.shot.getPosition().x < 0 ||
+                    s.shot.getPosition().x > DataContainer.mapGlobalSize.x ||
+                    s.shot.getPosition().y < 0 ||
+                    s.shot.getPosition().x > DataContainer.mapGlobalSize.y) {
                 s.shot.delete();
                 it.remove();
+                LL(this, "Deleting shot");
                 continue;
-//                LL(this, "Deleting shot");
             }
             for (Enemy enemy : enemies.getEnemies()) {
                 if (enemy.state.getCurrentState() == Enemy.DIYNG) {
@@ -131,10 +135,12 @@ public class Shooter {
 //                LL(this,"ECheking enemy");
 
                 if (enemy.getEnemyEntity().collision(s.shot.getPosition())) {
-                    enemy.doDamage(damage);
-                    s.shot.delete();
-                    it.remove();
-                    break;
+                    if(enemy.doDamage(damage)){
+                        s.shot.delete();
+                        LL(this, "Deleting shot");
+                        it.remove();
+                        break;
+                    }
                 }
             }
 //            LL(this, "updating  shot list");
